requirements.txt
streamlit
opencv-python
keras
py-sudoku
numpy
opencv-contrib-python